==> scripts/04_table_1.Rmd <==
---
title: "Untitled"
output: html_document
date: "2024-01-26"
---

```{r, include=FALSE}
library(tidyverse)
library(gtsummary)
library(sjlabelled)

df <- read_rds("../data/clean/01_clean_dataset.rds") %>%
  mutate(ratio = football_stadium_capacity / log2(municipality_population)) %>%
  var_labels(
    ratio = "Ratio of stadium capacity to log2(municipality population)"
  )

knitr::opts_chunk$set(echo = FALSE,
                      message = FALSE)
```

```{r}
df %>%
  select(conference, football_stadium_capacity, municipality_population, ratio, distance_nearest_emergency_department_miles, is_closest_ed_free_standing, acc_accreditation_or_certification, closest_ed_number_beds, teaching_hospital_in_municipality, drug_overdose_deaths, injury_deaths, primary_care_physicians, alcohol_impaired_driving_deaths, excessive_drinking) %>%
  tbl_summary(by = "conference") %>%
  add_overall()
```
