==> scripts/06_table_3.Rmd <==
---
title: "Tables of summary statistics by Kmeans cluster assignment"
output: word_document
date: "2024-01-26"
---

```{r, include=FALSE}
library(tidyverse)
library(gtsummary)

df <- read_rds("../data/clean/01_clean_dataset.rds")
clusters_3 <- read_rds("../output/03_clusters_3_no_ed_beds.rds")
clusters_4 <- read_rds("../output/03_clusters_4_no_ed_beds.rds")
clusters_5 <- read_rds("../output/03_clusters_5_no_ed_beds.rds")

knitr::opts_chunk$set(echo = FALSE,
                      message = FALSE)
```

```{r}
df %>%
  left_join(., clusters_3) %>%
  select(conference, football_stadium_capacity, municipality_population, distance_nearest_emergency_department_miles, is_closest_ed_free_standing, acc_accreditation_or_certification, closest_ed_number_beds, teaching_hospital_in_municipality, drug_overdose_deaths, injury_deaths, primary_care_physicians, alcohol_impaired_driving_deaths, excessive_drinking, cluster) %>%
  tbl_summary(by = "cluster")
```

```{r}
df %>%
  left_join(., clusters_4) %>%
  select(conference, football_stadium_capacity, municipality_population, distance_nearest_emergency_department_miles, is_closest_ed_free_standing, acc_accreditation_or_certification, closest_ed_number_beds, teaching_hospital_in_municipality, drug_overdose_deaths, injury_deaths, primary_care_physicians, alcohol_impaired_driving_deaths, excessive_drinking, cluster) %>%
  tbl_summary(by = "cluster")
```

Schools in each cluster:
```{r}
df %>%
  left_join(., clusters_4) %>%
  select(school, cluster) %>%
  arrange(cluster, school) %>%
  knitr::kable()
```


```{r}
df %>%
  left_join(., clusters_5) %>%
  select(conference, football_stadium_capacity, municipality_population, distance_nearest_emergency_department_miles, is_closest_ed_free_standing, acc_accreditation_or_certification, closest_ed_number_beds, teaching_hospital_in_municipality, drug_overdose_deaths, injury_deaths, primary_care_physicians, alcohol_impaired_driving_deaths, excessive_drinking, cluster) %>%
  tbl_summary(by = "cluster")
```
